requirements.txt
fastapi==0.95.0
uvicorn==0.21.0
tensorflow==2.11.0
pillow==9.4.0
numpy==1.23.5
pydantic==1.10.7
